// EmployeeManagement.Api/ClientApp/src/app/api/app.generated.ts
export interface DepartmentDto {
    id?: number;
    name?: string | undefined;
    abbreviation?: string | undefined;
    lastModAt?: Date;
    lastModBy?: string | undefined;
    createdAt?: Date;
    createdBy?: string | undefined;
}

export interface EmployeeCreateDto {
    name?: string | undefined;
    position?: string | undefined;
    phone?: string | undefined;
    username?: string | undefined;
    password?: string | undefined;
    bossId?: number | undefined;
    departmentId?: number;
}

export interface EmployeeDetailsDto {
    id?: number;
    name?: string | undefined;
    position?: string | undefined;
    phone?: string | undefined;
    departmentId?: number;
    departmentName?: string | undefined;
    lastModAt?: Date;
    lastModBy?: string | undefined;
    createdAt?: Date;
    createdBy?: string | undefined;
    username?: string | undefined;
    bossId?: number;
    bossName?: string | undefined;
}

export interface EmployeeListDto {
    id?: number;
    name?: string | undefined;
    position?: string | undefined;
    phone?: string | undefined;
    departmentId?: number;
    departmentName?: string | undefined;
    lastModAt?: Date;
    lastModBy?: string | undefined;
    createdAt?: Date;
    createdBy?: string | undefined;
}

export interface EmployeeUpdateDto {
    name?: string | undefined;
    position?: string | undefined;
    phone?: string | undefined;
    bossId?: number | undefined;
    departmentId?: number;
}